feat(top navigation): Handle status indicator
