+++ packages/pancake-uikit/src/components/MenuItem/styles.tsx
@@ -1,14 +1,28 @@
 import styled from "styled-components";
 import { MenuItemProps } from "./types";
 
-type StyledMenuItemProps = Pick<MenuItemProps, "isActive" | "variant">;
+type StyledMenuItemProps = Pick<MenuItemProps, "isActive" | "variant" | "statusColor">;
 
 const StyledMenuItem = styled.a<StyledMenuItemProps>`
+  position: relative;
   display: flex;
   align-items: center;
   color: ${({ theme, isActive }) => (isActive ? theme.menuItem.activeTextColor : theme.menuItem.textColor)};
   font-size: 16px;
   font-weight: ${({ isActive }) => (isActive ? "600" : "400")};
+
+  ${({ statusColor, theme }) =>
+    statusColor &&
+    `
+    &:after {
+      content: "";
+      border-radius: 100%;
+      background: ${theme.colors[statusColor]};
+      height: 8px;
+      width: 8px;
+      margin-left: 12px;
+    }
+  `}
 
   ${({ variant }) =>
     variant === "default"
@@ -32,7 +46,8 @@ const StyledMenuItem = styled.a<StyledMenuItemProps>`
 
   &:hover {
     background: ${({ theme }) => theme.menuItem.backgroundColor};
-    ${({ variant }) => variant === "default" && "border-radius: 16px;"}
+    ${({ variant }) => variant === "default" && "border-radius: 16px;"};
+  }
 `;
 
 export default StyledMenuItem;
