+++ packages/pancake-uikit/src/components/MenuItem/types.ts
@@ -1,9 +1,4 @@
-export type MenuItemTheme = {
-  textColor: string;
-  activeTextColor: string;
-  backgroundColor: string;
-  borderColor: string;
-};
+import { Colors } from "../../theme";
 
 export type MenuItemVariant = "default" | "subMenu";
 
@@ -11,4 +6,5 @@ export interface MenuItemProps {
   isActive?: boolean;
   href: string;
   variant?: MenuItemVariant;
+  statusColor?: keyof Colors;
 }
