+++ packages/pancake-uikit/src/theme/index.ts
@@ -1,6 +1,6 @@
 import { AlertTheme } from "../components/Alert/types";
 import { CardTheme } from "../components/Card/types";
-import { MenuItemTheme } from "../components/MenuItem/types";
+import { MenuItemTheme } from "../components/MenuItem/theme";
 import { SubMenuItemsTheme } from "../components/SubMenuItems/types";
 import { DropdownMenuItemTheme } from "../components/DropdownMenu/theme";
 import { PancakeToggleTheme } from "../components/PancakeToggle/types";
